re-arranged file structures and applied API fetch
