--- src/app/components/Row.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Row({ children, className, ...props }) {
-  return <div className={"flex flex-row w-full " + className}>{children}</div>;
-}
