+++ src/app/layout.js
@@ -1,50 +1,5 @@
-import { Inter } from "next/font/google";
+import { geist } from "@/config/fonts.config";
 import "./globals.css";
-import local from "next/font/local";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export const ligaSans = local({
-  src: [
-    {
-      path: "../assets/fonts/LigaSans-Bold.otf",
-      weight: "700",
-      style: "normal",
-    },
-  ],
-  variable: "--font-ligasans",
-});
-
-export const geist = local({
-  src: [
-    {
-      path: "../assets/fonts/Geist-Light.otf",
-      weight: "300",
-      style: "normal",
-    },
-    {
-      path: "../assets/fonts/Geist-Regular.otf",
-      weight: "400",
-      style: "normal",
-    },
-    {
-      path: "../assets/fonts/Geist-Medium.otf",
-      weight: "500",
-      style: "normal",
-    },
-    {
-      path: "../assets/fonts/Geist-SemiBold.otf",
-      weight: "600",
-      style: "normal",
-    },
-    {
-      path: "../assets/fonts/Geist-Bold.otf",
-      weight: "700",
-      style: "normal",
-    },
-  ],
-  variable: "--font-geist",
-});
 
 export const metadata = {
   title: "flex",
@@ -54,7 +9,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={inter.className + " min-h-full"}>{children}</body>
+      <body className={geist.className + " min-h-full"}>{children}</body>
     </html>
   );
 }
