+++ src/app/page.js
@@ -1,38 +1,17 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import { geist, ligaSans } from "./layout";
+import Container from "../components/Container";
+import Row from "../components/Row";
+import Column from "../components/Column";
 
-import Container from "./components/Container";
-import Row from "./components/Row";
-import Column from "./components/Column";
+import Footer from "../layouts/Footer";
+import Navbar from "../layouts/Navbar";
+import ProductSection from "@/sections/ProductSection";
 
-import Product from "./components/Product";
-
-import Footer from "./layouts/Footer";
-import Navbar from "./layouts/Navbar";
+import { ligaSans } from "@/config/fonts.config";
 
 export default function Home() {
-  const product = {
-    name: "Men's watch",
-    description: "Stylish men's watch for better lifestyle.",
-    link: {
-      label: "Buy now!",
-      href: "https://www.amazon.com",
-    },
-    imageUrl: "/watch2.png",
-  };
-
-  const product2 = {
-    name: "Men's wallet",
-    description: "Flex wallet on your pocket",
-    link: {
-      label: "Take a look!",
-      href: "https://www.amazon.com",
-    },
-    imageUrl: "/wallet.png",
-  };
-
   return (
     <>
       <Navbar />
@@ -49,7 +28,7 @@ export default function Home() {
 
             <Column>
               <Image
-                src={"/tshirt-top.png"}
+                src={"/man-top.png"}
                 alt="Man with t-shirt facing behind"
                 width={600}
                 height={400}
@@ -59,29 +38,7 @@ export default function Home() {
         </Container>
       </section>
 
-      <section className="py-8">
-        <Container>
-          <h1 className={ligaSans.className + " text-[2rem] mb-4"}>
-            Recommended Products
-          </h1>
-
-          <div className="grid-cols-2 grid gap-3">
-            <Product
-              name={product.name}
-              description={product.description}
-              link={product.link}
-              imageUrl={product.imageUrl}
-            />
-
-            <Product
-              name={product2.name}
-              description={product2.description}
-              link={product2.link}
-              imageUrl={product2.imageUrl}
-            />
-          </div>
-        </Container>
-      </section>
+      <ProductSection />
 
       <Footer />
     </>
