+++ src/components/Row.js
@@ -1,0 +1,3 @@
+export default function Row({ children, className = "", ...props }) {
+  return <div className={"flex flex-row w-full " + className}>{children}</div>;
+}
